+++ CRM/DataprocessorSearch/Form/MembershipSearch.php
@@ -111,15 +111,23 @@ class CRM_DataprocessorSearch_Form_MembershipSearch extends CRM_DataprocessorSea
    * Return altered rows
    *
    * Save the ids into the queryParams value. So that when an action is done on the selected record
-   * or on all records, the queryParams will hold all the activity ids so that in the next step only the selected record, or the first
-   * 50 records are populated.
-   *
-   * @param array $rows
-   * @param array $ids
-   *
+   * or on all records, the queryParams will hold all the activity ids so that in the next step only the selected record,
+   * or all records are populated.
    */
-  protected function alterRows(&$rows, $ids) {
-    $this->entityIDs = $ids;
+  protected function retrieveEntityIds() {
+    $this->dataProcessorClass->getDataFlow()->setLimit(false);
+    $this->dataProcessorClass->getDataFlow()->setOffset(0);
+    $this->entityIDs = [];
+    $id_field = $this->getIdFieldName();
+    try {
+      while($record = $this->dataProcessorClass->getDataFlow()->nextRecord()) {
+        if ($id_field && isset($record[$id_field])) {
+          $this->entityIDs[] = $record[$id_field]->rawValue;
+        }
+      }
+    } catch (\Civi\DataProcessor\DataFlow\EndOfFlowException $e) {
+      // Do nothing
+    }
     $this->_queryParams[0] = array(
       'membership_id',
       '=',