+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Refactored Field Output Handlers and simplified duplicate code for initializing a field.
 * Added Is not empty as an operator to the contact filter.
 * Added field to edit activity.
+* Renamed Case Role filter to Contact has role on case and added option to search for case without a role set.
 
 # Version 1.10.0